Moved login page redirect to user provider
