+++ src/context/UserContextProvider.tsx
@@ -1,6 +1,8 @@
-import { ReactNode, createContext } from 'react';
+import { ReactNode, createContext, useEffect } from 'react';
 import { User, UserContextType } from '@model';
+import { USER_STORAGE_KEY } from '@constants';
 import { useLocalStorage } from 'util/local-storage/useLocalStorage';
+import { useNavigate } from 'react-router-dom';
 
 interface Props {
   children: ReactNode | ReactNode[]
@@ -8,14 +10,20 @@ interface Props {
 
 export const UserContext = createContext<UserContextType | null>(null);
 
-const USER_STORAGE_KEY = 'agartex-user';
-
 const UserProvider = (props: Props) => {
+  const navigate = useNavigate();
+
   const { 
     storedValue: user, 
     setValue: setUser, 
     clearValue: logout
   } = useLocalStorage<User>(USER_STORAGE_KEY);
+
+  useEffect(() => {
+    if (!user) {
+      navigate('/login');
+    }
+  }, [user]);
 
   return (
     <UserContext.Provider
